requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
EasyProcess==0.3
entrypoint2==0.2.1
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mss==5.1.0
numpy==1.18.4
opencv-contrib-python==3.4.2.16
Pillow==7.1.2
pyscreenshot==2.1
PySimpleGUI==4.19.0
pystray==0.15.0
qrcode==6.1
requests==2.23.0
six==1.14.0
urllib3==1.25.9
Werkzeug==1.0.1
